+++ src/mopidy/audio/__init__.py
@@ -1,13 +1,10 @@
-from mopidy.types import PlaybackState
-
-from .actor import Audio, AudioProxy
-from .listener import AudioListener
-from .utils import supported_uri_schemes
+from mopidy.audio._actor import Audio, AudioProxy
+from mopidy.audio._listener import AudioListener
+from mopidy.audio._utils import supported_uri_schemes
 
 __all__ = [
     "Audio",
     "AudioListener",
     "AudioProxy",
-    "PlaybackState",  # Re-exported for backwards compatibility.
     "supported_uri_schemes",
 ]